test: separate prompt api contract integration coverage
